Renamed and deleted useless files
