+++ frontend/src/components/ui/FinancialFreedom/index.tsx
@@ -14,7 +14,7 @@ import MaskText from "@/components/Common/MaskText";
 import RevealCover from "@/components/Common/RevealCover";
 import { Div } from "../Featured/styles";
 import { imageVariants } from "../Featured";
-import { useIsMobile } from "../../../../libs/useIsMobile";
+import { useIsMobile } from "../../../lib/useIsMobile";
 import financial_freedom_banner from "../../../../public/images/financial_freedom_banner.png";
 import freedom_mobile_banner from "../../../../public/images/freedom_mobile_banner.png";
 import {
@@ -52,8 +52,7 @@ const FinancialFreedom = () => {
             variants={imageVariants}
             initial="hidden"
             whileInView="visible"
-            viewport={{ amount: 0.25, once: true }}
-          >
+            viewport={{ amount: 0.25, once: true }}>
             {isMobile ? (
               <Image src={freedom_mobile_banner} alt="banner_img" fill />
             ) : (
