+++ frontend/src/components/ui/OffersSection/index.tsx
@@ -1,5 +1,5 @@
-'use client';
-import Image from 'next/image';
+"use client";
+import Image from "next/image";
 import {
   Wrapper,
   Inner,
@@ -8,15 +8,15 @@ import {
   OfferCard,
   ImageCtn,
   TextCtn,
-} from './styles';
-import MaskText from '@/components/Common/MaskText';
-import { useIsMobile } from '../../../../libs/useIsMobile';
+} from "./styles";
+import MaskText from "@/components/Common/MaskText";
+import { useIsMobile } from "../../../lib/useIsMobile";
 import {
   desktopHeaderPhrases,
   desktopParagraphPhrase,
   mobileParagraphPhrase,
   offers,
-} from './constants';
+} from "./constants";
 
 const OffersSection = () => {
   const isMobile = useIsMobile();
