+++ frontend/src/lib/registry.tsx
