--- frontend/tailwind.config.js
@@ -1,14 +1,0 @@
-// tailwind.config.js
-// const {heroui} = require("@heroui/theme");
-import { heroui } from "@heroui/theme";
-/** @type {import('tailwindcss').Config} */
-module.exports = {
-  content: [
-    "./node_modules/@heroui/theme/dist/components/(button|ripple|spinner).js",
-  ],
-  theme: {
-    extend: {},
-  },
-  darkMode: "class",
-  plugins: [heroui()],
-};
